# rentflow-server/src/main/resources/openapi/rentflow-v1.yaml
openapi: 3.1.0
info:
  title: RentFlow API
  version: 1.0.0
  description: RentFlow V1 rental reservation API. All business timestamps include an offset and are stored as UTC.
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Notification page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
...
  /api/v1/notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Ulid'
      responses:
        '204':
          description: Notification marked read
  /api/v1/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Pending order created with a concrete equipment hold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '409':
          description: Quote, inventory, or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                reservationExpired:
                  value:
                    code: RESERVATION_EXPIRED
                    message: Reservation has expired
                    correlationId: 01J00000000000000000090004
                    details: {}
                reservationConsumed:
                  value:
                    code: RESERVATION_STATE_CONFLICT
                    message: Reservation cannot be consumed
                    correlationId: 01J00000000000000000090005
                    details: {}
    get:
      tags: [Orders]
      operationId: listOrders
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Current user orders sorted by createdAt DESC and orderId DESC
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
  /api/v1/orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      responses:
        '200':
          description: Current user order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
  /api/v1/orders/{orderId}/confirm:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [Orders]
      operationId: confirmOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Pending order confirmed, inventory hold consumed, and concrete equipment assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
  /api/v1/orders/{orderId}/receive:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [Orders]
      operationId: receiveOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Confirmed order received by its owner after the rental starts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
  /api/v1/orders/{orderId}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [Orders]
      operationId: cancelOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Pending order cancelled and inventory hold released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Ulid'
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Reuse only when retrying the same business intent.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[ -~]{16,128}$'
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Request validation failed
            correlationId: 01J00000000000000000090006
            details: {}
    AuthenticationRequired:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTHENTICATION_REQUIRED
            message: Authentication is required
            correlationId: 01J00000000000000000090007
            details: {}
    AccessDenied:
      description: The authenticated user does not own the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACCESS_DENIED
            message: Access is denied
            correlationId: 01J00000000000000000090008
            details: {}
...
    OrderNotFound:
      description: Order not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ORDER_NOT_FOUND
            message: Order was not found
            correlationId: 01J00000000000000000090011
            details: {}
    StateConflict:
      description: Resource state does not permit the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESERVATION_STATE_CONFLICT
            message: Reservation state does not permit this operation
            correlationId: 01J00000000000000000090012
            details: {}
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      example: 01J00000000000000000000001
    Timestamp:
      type: string
      format: date-time
      example: '2026-07-17T10:00:00Z'
    Money:
      type: string
      pattern: '^(0|[1-9][0-9]*)\.[0-9]{2}$'
      example: '200.00'
...
    PriceSnapshot:
      type: object
      additionalProperties: false
      required: [currency, pricingVersion, pricingRule, billingDays, dailyRate, rentalAmount, depositAmount, totalAmount, roundingMode]
      properties:
        currency:
          type: string
          const: CNY
        pricingVersion:
          type: integer
          format: int64
          minimum: 1
        pricingRule:
          type: string
          example: CEIL_24H_FIXED_DEPOSIT
        billingDays:
          type: integer
          minimum: 1
        dailyRate:
          $ref: '#/components/schemas/Money'
        rentalAmount:
          $ref: '#/components/schemas/Money'
        depositAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        roundingMode:
          type: string
          const: HALF_UP
...
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [quoteId]
      properties:
        quoteId:
          $ref: '#/components/schemas/Ulid'
    OrderStatus:
      type: string
      enum: [PENDING_CONFIRMATION, CONFIRMED, RECEIVED, CANCELLED, EXPIRED]
    Order:
      type: object
      additionalProperties: false
      required: [orderId, sourceReservationId, productId, productName, productModel, equipmentDisplayCode, status, effectiveStatus, startAt, endAt, expiresAt, priceSnapshot, createdAt, confirmedAt, receivedAt, cancelledAt, expiredAt]
      properties:
        orderId:
          $ref: '#/components/schemas/Ulid'
        sourceReservationId:
          $ref: '#/components/schemas/Ulid'
        productId:
          $ref: '#/components/schemas/Ulid'
        productName:
          type: string
        productModel:
          type: string
        equipmentDisplayCode:
          type: [string, 'null']
          description: Concrete unit assigned before dispatch; null until assignment.
        status:
          $ref: '#/components/schemas/OrderStatus'
        effectiveStatus:
          $ref: '#/components/schemas/OrderStatus'
        startAt:
          $ref: '#/components/schemas/Timestamp'
        endAt:
          $ref: '#/components/schemas/Timestamp'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        priceSnapshot:
          $ref: '#/components/schemas/PriceSnapshot'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        confirmedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        receivedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        cancelledAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        expiredAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
    OrderStatusHistory:
      type: object
      additionalProperties: false
      required: [toStatus, createdAt]
      properties:
        fromStatus:
          oneOf:
            - $ref: '#/components/schemas/OrderStatus'
            - type: 'null'
        toStatus:
          $ref: '#/components/schemas/OrderStatus'
        reason:
          type: [string, 'null']
        createdAt:
          $ref: '#/components/schemas/Timestamp'
    OrderDetail:
      type: object
      additionalProperties: false
      required: [orderId, sourceReservationId, productId, productName, productModel, equipmentDisplayCode, status, effectiveStatus, startAt, endAt, expiresAt, priceSnapshot, createdAt, confirmedAt, receivedAt, cancelledAt, expiredAt, statusHistory]
      properties:
        orderId:
          $ref: '#/components/schemas/Ulid'
        sourceReservationId:
          $ref: '#/components/schemas/Ulid'
        productId:
          $ref: '#/components/schemas/Ulid'
        productName:
          type: string
        productModel:
          type: string
        equipmentDisplayCode:
          type: [string, 'null']
          description: Concrete unit assigned before dispatch; null until assignment.
        status:
          $ref: '#/components/schemas/OrderStatus'
        effectiveStatus:
          $ref: '#/components/schemas/OrderStatus'
        startAt:
          $ref: '#/components/schemas/Timestamp'
        endAt:
          $ref: '#/components/schemas/Timestamp'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        priceSnapshot:
          $ref: '#/components/schemas/PriceSnapshot'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        confirmedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        receivedAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        cancelledAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        expiredAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/OrderStatusHistory'
    OrderPage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Notification:
      type: object
      additionalProperties: false
      required: [id, type, title, content, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        type:
          type: string
        title:
          type: string
        content:
          type: string
        readAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    NotificationPage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, details]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - ACCESS_DENIED
            - PRODUCT_NOT_FOUND
            - RESERVATION_NOT_FOUND
            - ORDER_NOT_FOUND
            - INVENTORY_NOT_AVAILABLE
            - RESERVATION_EXPIRED
            - RESERVATION_STATE_CONFLICT
            - RENTAL_ALREADY_STARTED
            - QUOTE_NOT_FOUND
            - QUOTE_EXPIRED
            - QUOTE_MISMATCH
            - QUOTE_ALREADY_CONSUMED
            - IDEMPOTENCY_CONFLICT
            - IDEMPOTENCY_IN_PROGRESS
            - ACTIVE_RESERVATION_LIMIT_EXCEEDED
            - ORDER_EXPIRED
            - ORDER_STATE_CONFLICT
            - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          $ref: '#/components/schemas/Ulid'
        details:
          type: object
          additionalProperties: true